% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skittles.R, R/skittles_raw.R
\docType{data}
\name{skittles}
\alias{skittles}
\alias{skittles_raw}
\title{Dataset of Skittles pack colour counts, with a pair of identical packs}
\format{
\subsection{skittles}{

A tibble with 468 rows and 7 columns:
\describe{
\item{pack_id}{A unique 3-character ID code for each row}
\item{strawberry, orange, lemon, apple, grape}{Skittle counts (integer)
per pack, for each colour/flavour}
\item{uncounted}{Number of uncounted fragments of possibly inedible
matter, or a chunk with no identifiable single colour. See the blog
post linked below for the detail of this fascinating methodological
decision.}
}
}

\subsection{skittles_raw}{

A data.frame with 468 rows and 6 columns:
\describe{
\item{strawberry, orange, lemon, apple, grape}{Skittle counts (integer)
per pack, for each colour/flavour}
\item{uncounted}{Number of uncounted fragments of possibly inedible
matter, or a chunk with no identifiable single colour. See the blog
post linked below for the detail of this fascinating methodological
decision.}
}
}
}
\source{
\url{https://github.com/possibly-wrong/skittles}
}
\usage{
skittles

skittles_raw
}
\description{
\subsection{skittles}{

Columns 2-7 are the original data from Possibly Wrong, with the column names
converted to lower case.
I added a "pack_id" key field as column 1, which is just the row number
zero-padded to a 3-digit character. (This is to avoid it being mistaken for
a numeric data field.)
No other data cleaning has been done; all the original columns read in
cleanly as integer vectors. This data is provided as a tibble rather than as
a plain data.frame.

This 'tidy' version of the dataset has also been exported to \verb{inst/} as
\code{skittles.csv}
}

\subsection{skittles_raw}{

The original data from Possibly Wrong is a tab-separated \code{.txt} file with
the field names in row 1 in title case, eg "Lemon".
The \code{skittles} dataset is a tidied version of this, but \code{skittles_raw}
retains the original field names and has no ID column added.
The numerical data is identical however.
This version of the dataset has also been exported to \verb{inst/} as
\code{skittles_raw.tsv}
}
}
\references{
\url{https://possiblywrong.wordpress.com/2019/04/06/follow-up-i-found-two-identical-packs-of-skittles-among-468-packs-with-a-total-of-27740-skittles/}
}
\keyword{datasets}
\keyword{skittles}
